Add caching to PRs, Files Changed, and also moved save_pr_data() into Database Manager
Because save_pr_data() is no longer a one-time initalizer function for the database, but can be run many times interacting with the database, it is now in Database Manager
