+++ src/database_init.py
@@ -46,6 +46,21 @@ def __init_ai_res_db():
                 )
     """
     )
+    cur.execute('''
+            CREATE UNIQUE INDEX IF NOT EXISTS "QuickFuncLookup" ON "functions" (
+                "classname"	ASC,
+                "function_name"	ASC
+            )
+            ''')
+    cur.execute(
+        """
+            CREATE TABLE IF NOT EXISTS "settings" (
+            "rowID"	INTEGER PRIMARY KEY AUTOINCREMENT,
+            "key"	TEXT UNIQUE NOT NULL,
+            "value"	TEXT NOT NULL
+        )
+    """
+    )
     cur.execute(
         """
             CREATE TABLE IF NOT EXISTS "settings" (
@@ -135,7 +150,7 @@ def __init_output_db():
         """
                       CREATE TABLE IF NOT EXISTS "pull_requests" (
                         "rowID"	INTEGER PRIMARY KEY AUTOINCREMENT,
-                        "pullNumber"	INTEGER NOT NULL,
+                        "pullNumber"	INTEGER UNIQUE NOT NULL,
                         "title"	TEXT NOT NULL,
                         "descriptionText"	TEXT,
                         "created"	TEXT NOT NULL,
@@ -183,6 +198,21 @@ def __init_output_db():
                 )
                 """
     )
+    cur.execute("""
+            CREATE UNIQUE INDEX IF NOT EXISTS "QuickFileClassFunction" ON "api_file_register" (
+                "filename"	ASC,
+                "commit_hash"	ASC,
+                "classname"	ASC,
+                "function_name"	ASC
+            );
+            """)
+    
+    cur.execute("""
+            CREATE UNIQUE INDEX IF NOT EXISTS "QuickFileChange" ON "files_changed" (
+                "filename"	ASC,
+                "commit_hash"	ASC
+            );
+            """)
 
     # Data structure for export... this defines the output table. It is an SQL view to use
     # all the power SQL has. query_generator generates the 1/0 columns.
@@ -288,224 +318,6 @@ def setup_caches():
     db.load_caches()
     db.close()
 
-
-def save_pr_data(pr_data: dict):
-    def clean_text(text):
-        """Replace newline characters with spaces, handling None values."""
-        if text is None:
-            return ""
-        return text.replace("\n", " ").replace("\r", " ")
-
-    def get_comments(data):
-        comments = data.get("comments")
-        clean_comments: list = []
-
-        if comments:
-            for comment in comments.values():
-                body = clean_text(comment.get("body", ""))
-                clean_comments.append(body)
-
-        return " | ".join(clean_comments)
-
-    def get_commits(data):
-        files_changed: list = []
-        clean_commits: list = []
-
-        commits = data.get("commits")
-
-        if commits:
-            for commit in commits.values():
-                commit_date = commit.get("date", "")
-                if commit_date:  # Only parse if commit_date is not empty
-                    try:
-                        commit_date_obj = datetime.strptime(
-                            commit_date, "%Y-%m-%dT%H:%M:%SZ"
-                        )
-                        clean_commits.append(
-                            (
-                                commit_date_obj,
-                                commit.get("sha", ""),
-                                commit.get("author_name", ""),
-                            )
-                        )
-                    except ValueError:
-                        continue  # Skip this commit if the date is invalid
-                files = commit.get("files", {})
-                if files:
-                    files_changed.extend(files.get("file_list", []))
-
-        # Sort commits by date (newest to oldest)
-        clean_commits.sort(key=lambda x: x[0], reverse=True)
-        sorted_commit_hashes = [commit[1] for commit in clean_commits]
-
-        return sorted_commit_hashes, files_changed
-
-    conn = sqlite3.connect(OUTPUT_DB)
-    cur = conn.cursor()
-
-    files_added = set()
-
-    # "issue": clean_text(pr.get("title", "")),
-    # "Pull Request": pr.get("is_pr", ""),
-    # "issue text": clean_text(pr.get("title", "")),
-    # "issue description": clean_text(pr.get("body", "")),
-    # "pull request text": clean_text(pr.get("title", "")),
-    # "pull request description": clean_text(pr.get("body", "")),
-    # "created_at": clean_text(pr.get("created_at", "")),
-    # "closed_at": clean_text(pr.get("closed_at", "")),
-    # "userlogin": clean_text(pr.get("userlogin", "")),
-
-    for num, data in pr_data.items():
-        is_pr = data["is_pr"]
-
-        print(f"Is {num} PR?: {is_pr}")
-
-        if not(is_pr):
-            continue # skip if not pr.
-
-        issue_num = num
-        issue_title = clean_text(data["title"])
-        issue_body = clean_text(data["body"])
-        created = clean_text(data["created_at"])
-        closed = clean_text(data["closed_at"])
-        userlogin = clean_text(data["userlogin"])
-        comments = clean_text(get_comments(data))
-
-        commit_hashes, files_changed = get_commits(data)
-        print(commit_hashes)
-        newest_commit_hash = commit_hashes[0]
-        author = newest_commit_hash[2]
-
-        commit_hashes = " | ".join(commit_hashes)
-
-        vals = (
-            issue_num,
-            issue_title,
-            issue_body,
-            created,
-            closed,
-            userlogin,
-            author,
-            newest_commit_hash,
-        )
-        cur.execute(
-            "INSERT INTO pull_requests (pullNumber, title, descriptionText, created, closed, userlogin, author, most_recent_commit) VALUES (?,?,?,?,?,?,?,?)",
-            vals,
-        )
-        for comment in comments:
-            if comment == "":
-                continue
-            cur.execute(
-                "INSERT INTO pull_request_comments (pullNumber, comment) VALUES (?,?)",
-                (issue_num, comment),
-            )
-
-        for file_change in set(files_changed):
-            if file_change == "":
-                continue
-            search = f"{newest_commit_hash}:{file_change}"
-            if search not in files_added:  # no duplicates!
-                cur.execute(
-                    "INSERT INTO files_changed (pullNumber, filename, commit_hash) VALUES (?, ?, ?)",
-                    (issue_num, file_change, newest_commit_hash),
-                )
-                files_added.add(search)
-            else:
-                cur.execute(
-                    "UPDATE files_changed SET pullNumber=? WHERE filename = ? AND commit_hash = ?",
-                    (issue_num, file_change, newest_commit_hash),
-                )
-        if newest_commit_hash != "":
-            for commit in commit_hashes:
-                cur.execute(
-                    "INSERT INTO pull_request_commits (pullNumber, commit_hash) VALUES (?,?)",
-                    (issue_num, commit),
-                )
-
-    conn.commit()
-    cur.close()
-    conn.close()
-
-
-def populate_db_with_mining_data():
-    """Go through the mining data "datamining.pkl" and add it to the database. Data import."""
-    conn = sqlite3.connect("./output/main.db")
-    cur = conn.cursor()
-
-    with open("./output/datamining.pkl", "rb") as file:
-        data = pickle.load(file)
-
-    files_added = set()
-
-    for row in tqdm.tqdm(data):
-        # Get elements from the rows
-        issue_num = row[1]
-        pr_check = row[2]
-        # If not a PR, skip loop
-        if pr_check is not True:
-            continue
-        issue_text = row[3]
-        issue_description = row[4]
-        created = row[7]
-        closed = row[8]
-        userlog = row[9]
-        author = row[10]
-        comments = row[11]
-        files_changed = row[12]
-        commit_hashes = row[13]
-        commit_hash = row[14]
-
-        # Insert into tables.
-        vals = (
-            issue_num,
-            issue_text,
-            issue_description,
-            created,
-            closed,
-            userlog,
-            author,
-            commit_hash,
-        )
-        cur.execute(
-            "INSERT INTO pull_requests (pullNumber, title, descriptionText, created, closed, userlogin, author, most_recent_commit) VALUES (?,?,?,?,?,?,?,?)",
-            vals,
-        )
-        for comment in comments:
-            if comment == "":
-                continue
-            cur.execute(
-                "INSERT INTO pull_request_comments (pullNumber, comment) VALUES (?,?)",
-                (issue_num, comment),
-            )
-        for file_change in set(files_changed):
-            if file_change == "":
-                continue
-            search = f"{commit_hash}:{file_change}"
-            if search not in files_added:  # no duplicates!
-                cur.execute(
-                    "INSERT INTO files_changed (pullNumber, filename, commit_hash) VALUES (?, ?, ?)",
-                    (issue_num, file_change, commit_hash),
-                )
-                files_added.add(search)
-            else:
-                cur.execute(
-                    "UPDATE files_changed SET pullNumber=? WHERE filename = ? AND commit_hash = ?",
-                    (issue_num, file_change, commit_hash),
-                )
-        if commit_hash != "":
-            for commit in commit_hashes:
-                cur.execute(
-                    "INSERT INTO pull_request_commits (pullNumber, commit_hash) VALUES (?,?)",
-                    (issue_num, commit),
-                )
-
-    cur.execute(
-        "INSERT INTO settings (key, value) VALUES ('setup', ?)",
-        (datetime.now(),),
-    )
-    conn.commit()
-    cur.close()
-    conn.close()
 
 
 def query_generator():
